Move FiniteStateMachine.js into ai/
It's only imported by FighterFSM.js and ShuttleFSM.js - nothing else in
the codebase uses the generic FSM engine, so it belongs next to its only
consumers rather than in the utils/ grab-bag.
